requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
pydantic==2.8.2
numpy==1.26.4
python-multipart==0.0.9
deepface==0.0.93
motor==3.5.1
pymongo==4.8.0
Pillow==10.4.0
tensorflow==2.17.0
opencv-python==4.10.0.84
matplotlib==3.9.2
python-dotenv==1.0.1
gunicorn==23.0.0 
mtcnn
tf-keras